refactor: split cutter responsibilities (#140)
